@@ -1,9 +1,10 @@
 module github.com/botherder/androidqf
 
-go 1.18
+go 1.20
 
 require (
-	filippo.io/age v1.0.0
+	filippo.io/age v1.1.1
+	github.com/avast/apkverifier v0.0.0-20221110131049-7720fc1ebef0
 	github.com/botherder/go-savetime v1.4.0
 	github.com/i582/cfmt v1.4.0
 	github.com/manifoldco/promptui v0.9.0
@@ -11,9 +12,11 @@ require (
 )
 
 require (
+	github.com/avast/apkparser v0.0.0-20221012080151-bfc57d4d0502 // indirect
 	github.com/chzyer/readline v0.0.0-20180603132655-2972be24d48e // indirect
 	github.com/gookit/color v1.3.2 // indirect
-	golang.org/x/crypto v0.0.0-20210921155107-089bfa567519 // indirect
-	golang.org/x/sys v0.0.0-20220919091848-fb04ddd9f9c8 // indirect
+	github.com/klauspost/compress v1.15.12 // indirect
+	golang.org/x/crypto v0.4.0 // indirect
+	golang.org/x/sys v0.3.0 // indirect
 	gopkg.in/check.v1 v1.0.0-20201130134442-10cb98267c6c // indirect
 )
