@@ -9,6 +9,8 @@ import (
 	"fmt"
 	"reflect"
 	"strings"
+    "strconv"
+    "github.com/avast/apkverifier"
 )
 
 type PackageFile struct {
@@ -18,6 +20,11 @@ type PackageFile struct {
 	SHA1      string `json:"sha1"`
 	SHA256    string `json:"sha256"`
 	SHA512    string `json:"sha512"`
+    Error               string               `json:"error"`
+    VerifiedCertificate bool                 `json:"verified_certificate"`
+    Certificate         apkverifier.CertInfo `json:"certificate"`
+    CertificateError    string               `json:"certificate_error"`
+    TrustedCertificate  bool                 `json:"trusted_certificate"`
 }
 
 type Package struct {
@@ -30,60 +37,35 @@ type Package struct {
 	ThirdParty bool          `json:"third_party"`
 }
 
-func (a *ADB) getPackageFiles(packageName string) []PackageFile {
-	out, err := a.Shell("pm", "path", packageName)
-	if err != nil {
-		fmt.Printf("Failed to get file paths for package %s: %v: %s\n", packageName, err, out)
-		return []PackageFile{}
-	}
-
-	packageFiles := []PackageFile{}
-	for _, line := range strings.Split(out, "\n") {
-		packagePath := strings.TrimPrefix(strings.TrimSpace(line), "package:")
-		if packagePath == "" {
-			continue
-		}
-
-		packageFile := PackageFile{
-			Path: packagePath,
-		}
-
-		md5Out, err := a.Shell("md5sum", packagePath)
-		if err == nil {
-			packageFile.MD5 = strings.SplitN(md5Out, " ", 2)[0]
-		}
-		sha1Out, err := a.Shell("sha1sum", packagePath)
-		if err == nil {
-			packageFile.SHA1 = strings.SplitN(sha1Out, " ", 2)[0]
-		}
-		sha256Out, err := a.Shell("sha256sum", packagePath)
-		if err == nil {
-			packageFile.SHA256 = strings.SplitN(sha256Out, " ", 2)[0]
-		}
-		sha512Out, err := a.Shell("sha512sum", packagePath)
-		if err == nil {
-			packageFile.SHA512 = strings.SplitN(sha512Out, " ", 2)[0]
-		}
-
-		packageFiles = append(packageFiles, packageFile)
-	}
-
-	return packageFiles
-}
 
 // GetPackages returns the list of installed package names.
 func (a *ADB) GetPackages() ([]Package, error) {
-	out, err := a.Shell("pm", "list", "packages", "-u", "-i")
-	if err != nil && out == "" {
-		return []Package{}, fmt.Errorf("failed to launch `pm list packages` command: %v",
-			err)
-	}
+    withInstaller := true
+	out, err := a.Shell("pm", "list", "packages", "-U", "-u", "-i")
+
+    if err != nil {
+        // Some phones do not support -i option
+        out, err = a.Shell("pm", "list", "packages", "-U", "-u")
+        if err != nil {
+            return []Package{}, fmt.Errorf("failed to launch `pm list packages` command: %v",
+                err)
+        }
+        withInstaller = false
+    }
 
 	packages := []Package{}
+    var installer string
+    var uid int
 	for _, line := range strings.Split(out, "\n") {
 		fields := strings.Fields(line)
 		packageName := strings.TrimPrefix(strings.TrimSpace(fields[0]), "package:")
-		installer := strings.TrimPrefix(strings.TrimSpace(fields[1]), "installer=")
+        if withInstaller {
+        installer = strings.TrimPrefix(strings.TrimSpace(fields[1]), "installer=")
+            uid, _ = strconv.Atoi(strings.TrimPrefix(strings.TrimSpace(fields[2]), "uid:"))
+        } else {
+            uid, _ = strconv.Atoi(strings.TrimPrefix(strings.TrimSpace(fields[1]), "uid:"))
+            installer = ""
+        }
 
 		if packageName == "" {
 			continue
@@ -92,10 +74,11 @@ func (a *ADB) GetPackages() ([]Package, error) {
 		newPackage := Package{
 			Name:       packageName,
 			Installer:  installer,
+            UID:        uid,
 			Disabled:   false,
 			System:     false,
 			ThirdParty: false,
-			Files:      a.getPackageFiles(packageName),
+            Files:      []PackageFile{},
 		}
 
 		packages = append(packages, newPackage)
@@ -137,4 +120,26 @@ func (a *ADB) GetPackages() ([]Package, error) {
 	}
 
 	return packages, nil
+}
+
+// GetPackagePaths returns a list of file paths associated with the provided
+// package name.
+func (a *ADB) GetPackagePaths(packageName string) ([]string, error) {
+    out, err := a.Shell("pm", "path", packageName)
+    if err != nil {
+        return []string{}, fmt.Errorf("failed to launch `pm path` command: %v",
+            err)
+    }
+
+    packagePaths := []string{}
+    for _, line := range strings.Split(out, "\n") {
+        packagePath := strings.TrimPrefix(strings.TrimSpace(line), "package:")
+        if packagePath == "" {
+            continue
+        }
+
+        packagePaths = append(packagePaths, packagePath)
+    }
+
+    return packagePaths, nil
 }
